+++ CHANGELOG.md
@@ -11,6 +11,7 @@ for further information about branching and tagging conventions.
 - Abbreviation for balances
 - General KYC form
 - Display first name and last name customer
+- Buying atomic swaps
 
 #### Removed
 - Asset codes from drawer
@@ -20,6 +21,7 @@ for further information about branching and tagging conventions.
 - Bug with validation on `Recipient email or phone` field
 - Bug when user can to create atomic swap for not own asset
 - Bug when user cannot send money
+- Bug when the login form is cleared and enabled before routing to the app
 
 #### Changed
 - `My businesses` default tab